Consolidate blog related functionality
