+++ lib/index.ts
@@ -1,4 +1,5 @@
 export { colors } from './colors';
+export { formatDate } from './date';
 export { MDX_COMPONENTS } from './mdx';
 export { useSeo } from './seo';
 export { isExternal } from './url';
